OK-119 : Migrate more ilmoittautuminen data
Logentries of ilmoittautuminen can be found
at least in three different forms in
Valintatulos documents:

1) in the most recent ones, there's an explicit
"ilmoittautumisTila" marker in the logentry "muutos"
field

2) in some cases, there's "ILMOITETTU" in "muutos" field

3) and sometimes, maybe in the old 2ndary degree cases,
there's just "VASTAANOTTANUT" because this is typically
saved with ilmoittautuminen on secondary degree.

If there's a "non-empty" data on ilmoittautuminen (!== "EI_TEHTY"),
but none of these logentries match, let's set epoch as timestamp.

If there's just the Valintatulos default value of "EI_TEHTY" and
no matching logentries, nothing needs to be saved. Ilmoittautuminen
is anyway relevant only with vastaanotto.