Change candidate polling query join logic
The viestinnan_ohjaus table is preliminary empty, because no
data migration has been done for the data. This means that
we can't use it as start table when joining with valinnantulos
and hakukohde tables, because the query would never return
even a single row, unless there was a row which links to a
valinnantulos row. Now the query returns candidates and if
there is no row for the candidate in viestinnan_ohjaus
table, then vo.sent will be null.