BUG-1758 : Add a more performant API
kkhakijat API of Suoritusrekisteri retrieves the latest
lukuvuosimaksu data of each hakukohde of each hakutoive of
each person of a hakukohde. When people have several
hakutoive this can lead to lots and lots of queries and in
the end there is very little data to return to so it makes
sense to retrieve it all at once.